Schematic Specification Version 2: Introduces Biomes and Entity Support.
The second version of the Sponge Schematic Format. All changes have been
considered by multiple community members as a means to easily upgrade
between Version 1 and Version 2. Notably, due to the introduction of
Minecraft's Data Version, we can now safely "upgrade" schematics to
newer Minecraft versions without relying on upgrading the Schematic
itself.

The summary of changes are:
- Change `TileEntities` to `BlockEntities`
- Addition of Entity object support
- Replacement of ContentVersion with one Data Version to use Minecraft's
DataFixers. This means that a Schematic does not have to be upgraded,
but it can be applied to various Minecraft world instances as long as
the manager of those world instances utilizes the DataFixers for the
appropriate DataVersion.
- Addition of Biomes at coordinates. Specifically Biomes are not
supported in a 3D space, they are still applied in 2D coordinates.
Potential future changes to Minecraft may introduce Bimoes at the 3D
level, but when that occurs, we will adjust the schematic format
accordingly.
- Clarification of NBT type usages for various fields and object types.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>